Move jsonschema and paths to models
