+++ pkg/models/jsonschema/gallery.go
