+++ pkg/models/jsonschema/image.go
