+++ pkg/models/jsonschema/mappings.go
