+++ pkg/models/jsonschema/scene.go
