+++ pkg/models/jsonschema/scraped.go
