+++ pkg/models/jsonschema/utils.go
