+++ pkg/models/paths/paths_generated.go
