+++ pkg/models/paths/paths_scene_markers.go
