+++ pkg/models/paths/paths_scenes.go
