+++ pkg/performer/export_test.go
@@ -4,8 +4,8 @@ import (
 	"database/sql"
 	"errors"
 
-	"github.com/stashapp/stash/pkg/manager/jsonschema"
 	"github.com/stashapp/stash/pkg/models"
+	"github.com/stashapp/stash/pkg/models/jsonschema"
 	"github.com/stashapp/stash/pkg/models/mocks"
 	"github.com/stashapp/stash/pkg/utils"
 	"github.com/stretchr/testify/assert"
